Conic constriants, knotpoints::AbstractVector, DiscreteTrajectory (#6)
Knotpoints are now a single AbstractVector and the knotpoint function backend pass in views of this vector to the function implementations. This plays nicely with ForwardDiff and functions that span multiple knotpoints. The DiscreteTrajectory struct is introduced to hold onto this knotpoint vector along with other info like timesteps. The DiscreteTrajectory can produce self views which is what is passed into the knot point implementation functions. Conic form constraints are now used for even tighter integration with Clarabel.